requirements.txt
Flask==3.0.0
requests==2.25.1
gunicorn
python-dotenv==0.21.1
Werkzeug==0.16.1